Add MapGenTest.
