+++ OpenKh.Tests/kh2/res/mapgen/multi-4bpp/multi-4bpp.doct.dump
@@ -1,0 +1,3 @@
+[Entry1] Unk=0; BoundingBox=Minimum: <-1514, -917, -1514> Maximum: <1514, 489, 1514>
+ [Entry1] Unk=0; BoundingBox=Minimum: <-1514, -917, -1514> Maximum: <1514, 489, 1514>
+  [Entry2] Flags=0; BoundingBox=Minimum: <-1514, -917, -1514> Maximum: <1514, 489, 1514>
